Add api routing Add make run
