+++ Makefile
@@ -13,7 +13,12 @@ install:
 	$(PIP) install --upgrade pip
 	$(PIP) install -r requirements.txt
 
+run:
+	@echo "Running the Flask application..."
+	$(PYTHON) $(APP)
+
 help:
 	@echo "Makefile Commands:"
 	@echo "  venv		 Create a virtual environment"
 	@echo "  install     Install dependencies from requirements.txt"
+	@echo "  run         Run the Flask application"
